Added filtersLogicalOperator to Sensor filter dependency - Improved Sensor filter dependency tests
